[f] 213 - Refactor code
https://www.pivotaltracker.com/story/show/187724213

+++ app/interactors/billing/reservations/save_room_reservation.rb
@@ -19,6 +19,6 @@ class Billing::Reservations::SaveRoomReservation
   end
 
   def rollback
-    context.reservation.destroy
+    context.reservation.destroy if context.reservation&.persisted?
   end
 end
